Component 2 V4
Created v4 of my 2nd component.

A while loop has been set up to solve the issue where after an invalid input has been inputted, it quit the program instead of asking it again.

Boundary values have also been set up to make sure that if the user inouts an integer larger than 1 or less than 2, then it will tell them that it is not acceptable and asks them to choose between those two integers instead.

ValueError has been added to ensure that the input is not a letter, but rather stays as an integer instead.