component 10 v2
created an int validation fucntion that checks if the input is either between boundary values, or an invalid input (other non-numerical characters).

allows boundary values to be modified using the low and high variables, which by default in this program have been set to 1 as the low, and 2 as the high.